db presque fini mais fonctionne

+++ app/src/main/java/com/example/db/models/Personne.java
@@ -1,7 +1,7 @@
 package com.example.db.models;
 
 public class Personne {
-    private int id;
+    private long id;
     private String nom;
     private String mdp;
 
@@ -11,7 +11,7 @@ public class Personne {
         this.mdp = mdp;
     }
     // Avec l'id
-    public Personne(int id, String nom, String mdp) {
+    public Personne(long id, String nom, String mdp) {
         this.id = id;
         this.nom = nom;
         this.mdp = mdp;
@@ -19,7 +19,7 @@ public class Personne {
 
     // Getters and setters
 
-    public int getId() { return id; }
+    public long getId() { return id; }
     public void setId(int id) { this.id = id; }
     public String getNom() { return nom; }
     public void setNom(String nom) { this.nom = nom; }
